Directions For Tlauncher
Follow the installation process, but when you have the app on your desktop, right-click it and select the properties option. Find the unblock option at the bottom of the window and click it. Then hit apply and ok. Do this for the setup download if it won't run the setup. Regardless after you need right-click the new file on your desktop and select the properties option. Find the unblock option at the bottom of the window and click it. Then hit apply and ok. If this does not work, right-click the game on the desktop and click the "open file location" option. Then click the new highlighted application and go to properties and unblock it. Then launch the setup from there. Unblocking your apps in the properties window is needed for all downloaded apps that aren't from the Microsoft store. It is rare that your computer allows you to run files from the internet without it being unblocked.